+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## [0.1.8] - 30/07/2020.
+## [0.1.9] - 30/07/2020.
 
-* Update Dependency
-* add documentation
+* placeholder for cachedUrl
+* errorWidget used if there are problem